+++ src/main/java/com/tristankechlo/livingthings/init/ModEntityTypes.java
@@ -8,6 +8,7 @@ import com.tristankechlo.livingthings.entities.LionEntity;
 import com.tristankechlo.livingthings.entities.MantarayEntity;
 import com.tristankechlo.livingthings.entities.OstrichEntity;
 import com.tristankechlo.livingthings.entities.PenguinEntity;
+import com.tristankechlo.livingthings.entities.RaccoonEntity;
 import com.tristankechlo.livingthings.entities.SharkEntity;
 
 import java.util.List;
@@ -43,6 +44,7 @@ public class ModEntityTypes {
     public static final EntityType<FlamingoEntity> FLAMINGO_ENTIY = createEntity("flamingo", FlamingoEntity::new, EntityClassification.CREATURE, 0.6F, 1.25F, 0xf38989, 0x2d0404);
     public static final EntityType<CrabEntity> CRAB_ENTITY = createEntity("crab", CrabEntity::new, EntityClassification.CREATURE, 0.4F, 0.4F, 0xeb4034, 0x73706f);
     public static final EntityType<MantarayEntity> MANTARAY_ENTITY = createEntity("mantaray", MantarayEntity::new, EntityClassification.WATER_AMBIENT, 0.75F, 0.45F, 0x000896, 0x595a6b);
+    public static final EntityType<RaccoonEntity> RACCOON_ENTITY = createEntity("raccoon", RaccoonEntity::new, EntityClassification.CREATURE, 0.5F, 0.75F, 0x6e6e6e, 0x000000);
 
 
 	/**
@@ -58,6 +60,7 @@ public class ModEntityTypes {
 		GlobalEntityTypeAttributes.put(FLAMINGO_ENTIY, FlamingoEntity.getAttributes().create());
 		GlobalEntityTypeAttributes.put(CRAB_ENTITY, CrabEntity.getAttributes().create());
 		GlobalEntityTypeAttributes.put(MANTARAY_ENTITY, MantarayEntity.getAttributes().create());
+		GlobalEntityTypeAttributes.put(RACCOON_ENTITY, RaccoonEntity.getAttributes().create());
 	}
 
 	/**
@@ -73,6 +76,7 @@ public class ModEntityTypes {
 		EntitySpawnPlacementRegistry.register(FLAMINGO_ENTIY, PlacementType.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, AnimalEntity::canAnimalSpawn);
 		EntitySpawnPlacementRegistry.register(CRAB_ENTITY, PlacementType.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, CrabEntity::canCrabSpawn);
 		EntitySpawnPlacementRegistry.register(MANTARAY_ENTITY, PlacementType.IN_WATER, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, MantarayEntity::canMantaraySpawn);
+		EntitySpawnPlacementRegistry.register(RACCOON_ENTITY, PlacementType.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, AnimalEntity::canAnimalSpawn);
 	}
 
     /**