+++ src/main/java/com/tristankechlo/livingthings/init/ModSounds.java
@@ -30,6 +30,10 @@ public class ModSounds {
 	public static final RegistryObject<SoundEvent> OSTRICH_EGG_CRACKS = SOUNDS.register("ostrich_egg_cracks", () -> new SoundEvent(getSound("ostrich.egg.cracks")));
 	public static final RegistryObject<SoundEvent> OSTRICH_EGG_LAYING = SOUNDS.register("ostrich_egg_laying", () -> new SoundEvent(getSound("ostrich.egg.laying")));
 	public static final RegistryObject<SoundEvent> OSTRICH_AMBIENT = SOUNDS.register("ostrich_ambient", () -> new SoundEvent(getSound("ostrich.ambient")));
+
+	public static final RegistryObject<SoundEvent> RACCOON_AMBIENT = SOUNDS.register("raccoon_ambient", () -> new SoundEvent(getSound("raccoon.ambient")));
+	public static final RegistryObject<SoundEvent> RACCOON_HURT = SOUNDS.register("raccoon_hurt", () -> new SoundEvent(getSound("raccoon.hurt")));
+	public static final RegistryObject<SoundEvent> RACCOON_DEATH = SOUNDS.register("raccoon_death", () -> new SoundEvent(getSound("raccoon.death")));
 	
 	
 	private static ResourceLocation getSound(String name) {